refactor(traversal): unify divergent TraversalOptions structs (#58)
Four TraversalOptions-family structs existed across the workspace. This
PR eliminates the two non-canonical ones by pointing all live-path code
at khive-storage::TraversalOptions (the canonical S1) and deleting the
isolated legacy module.

Changes by struct:
- S1 (khive-storage): Replace #[derive(Default)] with a custom impl
  setting max_depth=3 (consensus default from S2/S3; was 0 via
  usize::default()). Add regression test asserting ::default().max_depth==3.
- S2 (khive-runtime::graph_traversal): Delete the duplicate struct + its
  custom Default impl. Add TraversalOptions to the khive_storage::types
  use import. Rename options.max_results -> options.limit with u32->usize
  cast. Fix khive-runtime/src/lib.rs re-export to use khive_storage::types.
- S3 (khive-retrieval::graph, feature-gated graph-legacy): Delete the
  entire crates/khive-retrieval/src/graph/ directory (8 files). Remove
  the cfg(feature="graph-legacy") blocks from lib.rs and the feature
  declaration from Cargo.toml. Zero callers outside the module confirmed
  before deletion.
- S4 (khive-pack-kg TraverseParams wire deserializer): Unchanged.

call-site audit: grep across workspace for TraversalOptions::default()
and ..Default::default() — all callers that construct TraversalOptions
set max_depth explicitly. The Default change affects no existing behavior.
S3 zero-caller confirmation: no crate outside khive-retrieval/src/graph/
referenced graph-legacy or khive_retrieval::graph.
